feat: footer refactor: Move components to module folder
