+++ src/app/(home)/contact/page.tsx
@@ -1,3 +1,5 @@
-export const Page = () => {
+const Page = () => {
 	return <div>Hello</div>;
 };
+
+export default Page;
