+++ src/app/(home)/layout.tsx
@@ -1,4 +1,5 @@
-import { Navbar } from "@/app/(home)/navbar";
+import { Footer } from "@/module/home/footer";
+import { Navbar } from "@/module/home/navbar";
 
 type Props = {
 	children: React.ReactNode;
@@ -7,7 +8,8 @@ const Layout = ({ children }: Props) => {
 	return (
 		<div className="flex min-h-screen flex-col">
 			<Navbar />
-			{children}
+			<div className="flex-1 bg-[#f4f4f4]">{children}</div>
+			<Footer />
 		</div>
 	);
 };
