+++ src/components/ui/progress.tsx
@@ -16,7 +16,7 @@ function Progress({
       className={cn(
         "relative h-3 w-full overflow-hidden rounded-full",
         // Modified
-        "border bg-transparent",
+        "border bg-white",
         className
       )}
       {...props}
