+++ src/module/home/footer.tsx
@@ -1,0 +1,9 @@
+export const Footer = () => {
+	return (
+		<footer className="flex justify-between border-t p-6 font-medium">
+			<div className="flex items-center gap-2">
+				<p>vcom, Inc.</p>
+			</div>
+		</footer>
+	);
+};
