+++ src/module/home/navbar.tsx
@@ -5,9 +5,9 @@ import { Poppins } from "next/font/google";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { useState } from "react";
-import { NavbarSidebar } from "@/app/(home)/navbar-sidebar";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
+import { NavbarSidebar } from "@/module/home/navbar-sidebar";
 
 const poppins = Poppins({
 	subsets: ["latin"],
